Add Business and Campaigns Pages with Loading States and Dialogs
Introduce a new BusinessesPage component for managing business profiles, including a create business dialog with form validation. Implement a CampaignsPage component for managing broadcast campaigns, featuring search and filter functionality, as well as a create campaign dialog. Enhance user experience with loading states and error handling for both pages, ensuring a seamless interaction when managing businesses and campaigns.
